requirements.txt
python-telegram-bot>=13.0
requests
pytz
numexpr
apscheduler
scikit-learn
nltk
numpy
